DisplayLang読み込みをcomposableに統一し、POINTSテーブルを分離、不要なevent.jsを削除
