+++ src/popup/NagaList.vue
@@ -42,6 +42,8 @@
 import { onMounted, reactive, ref, computed } from "vue";
 import Kyoku from "@/popup/Kyoku.vue";
 import { extractTable, toSoulTable, toNagaLog, fixScoreRonTileWasReachTile } from "@/lib/naga";
+import { POINTS } from "@/lib/points";
+import { useDisplayLang } from "@/composables/useDisplayLang";
 
 export default {
   components: { Kyoku },
@@ -184,14 +186,7 @@ export default {
       }
     }
 
-    //表示用言語の設定をlocal storageから呼び出す
-    const DisplayLang = ref(0)
-    chrome.storage.local.get("DisplayLang", (result) => {
-      // join langs
-      if (typeof result.DisplayLang !== "undefined") {
-        DisplayLang.value = result.DisplayLang;
-      }
-    });
+    const DisplayLang = useDisplayLang();
 
     //content-scriptに通信して牌譜を送信させる
     onMounted(() => {
@@ -373,94 +368,6 @@ export default {
       return JSON.parse(JSON.stringify(src));
     }
 
-    // 段位ごとのポイント配分
-    const POINTS = {
-      east: {
-        bronze: {
-          "初心★1": [25, 10, -5, -15],
-          "初心★2": [25, 10, -5, -15],
-          "初心★3": [25, 10, -5, -15],
-          "雀士★1": [25, 10, -5, -25],
-          "雀士★2": [25, 10, -5, -35],
-          "雀士★3": [25, 10, -5, -45]
-        },
-        silver: {
-          "雀士★1": [35, 15, -5, -25],
-          "雀士★2": [35, 15, -5, -35],
-          "雀士★3": [35, 15, -5, -45],
-          "雀傑★1": [35, 15, -5, -55],
-          "雀傑★2": [35, 15, -5, -65],
-          "雀傑★3": [35, 15, -5, -75]
-        },
-        gold: {
-          "雀傑★1": [55, 25, -5, -55],
-          "雀傑★2": [55, 25, -5, -65],
-          "雀傑★3": [55, 25, -5, -75],
-          "雀豪★1": [55, 25, -5, -95],
-          "雀豪★2": [55, 25, -5, -105],
-          "雀豪★3": [55, 25, -5, -115]
-        },
-        tama: {
-          "雀豪★1": [70, 35, -5, -95],
-          "雀豪★2": [70, 35, -5, -105],
-          "雀豪★3": [70, 35, -5, -115],
-          "雀聖★1": [70, 35, -5, -125],
-          "雀聖★2": [70, 35, -5, -135],
-          "雀聖★3": [70, 35, -5, -145]
-        },
-        king: {
-          "雀聖★1": [75, 35, -5, -125],
-          "雀聖★2": [75, 35, -5, -135],
-          "雀聖★3": [75, 35, -5, -145],
-        }
-      },
-      south: {
-        bronze: {
-          "初心★1": [35, 15, -5, -15],
-          "初心★2": [35, 15, -5, -15],
-          "初心★3": [35, 15, -5, -15],
-          "雀士★1": [35, 15, -5, -35],
-          "雀士★2": [35, 15, -5, -55],
-          "雀士★3": [35, 15, -5, -75]
-        },
-        silver: {
-          "雀士★1": [55, 25, -5, -35],
-          "雀士★2": [55, 25, -5, -55],
-          "雀士★3": [55, 25, -5, -75],
-          "雀傑★1": [55, 25, -5, -95],
-          "雀傑★2": [55, 25, -5, -115],
-          "雀傑★3": [55, 25, -5, -135]
-        },
-        gold: {
-          "雀傑★1": [95, 45, -5, -95],
-          "雀傑★2": [95, 45, -5, -115],
-          "雀傑★3": [95, 45, -5, -135],
-          "雀豪★1": [95, 45, -5, -180],
-          "雀豪★2": [95, 45, -5, -195],
-          "雀豪★3": [95, 45, -5, -210]
-        },
-        tama: {
-          "雀豪★1": [125, 60, -5, -180],
-          "雀豪★2": [125, 60, -5, -195],
-          "雀豪★3": [125, 60, -5, -210],
-          "雀聖★1": [125, 60, -5, -225],
-          "雀聖★2": [125, 60, -5, -240],
-          "雀聖★3": [125, 60, -5, -255]
-        },
-        king: {
-          "雀聖★1": [135, 65, -5, -225],
-          "雀聖★2": [135, 65, -5, -240],
-          "雀聖★3": [135, 65, -5, -255],
-        }
-      },
-      others: {
-        1030: [50, 10, -10, -30],
-        1020: [40, 10, -10, -20],
-        515: [35, 5, -5, -15],
-        510: [30, 5, -5, -10],
-        tenho: [90, 45, 0, -135],
-      }
-    }
     return {
       Kyoku_info,
       select,
